requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.6.0+cu124
torchvision==0.21.0+cu124

accelerate==1.13.0
datasets==4.8.2
Faker==40.11.0
numpy==2.4.3
opencv_python_headless==4.10.0.84
pdf2image==1.17.0
Pillow==12.1.1
python-dotenv==1.2.2
PyMuPDF==1.26.1
transformers==5.3.0